pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reliapi"
version = "1.1.0"
description = "Reliability layer for API calls: retries, caching, dedup, circuit breakers, LLM proxy with cost control."
readme = "README.md"
license = "AGPL-3.0-only"
license-files = ["LICENSE"]
authors = [
    {name = "KikuAI-Lab", email = "dev@kikuai.dev"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "PyYAML>=6.0.1",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
    "black>=23.12.0",
    "mypy>=1.8.0",
]

[project.urls]
"Homepage" = "https://github.com/kiku-jw/reliapi"
"Documentation" = "https://reliapi.kikuai.dev"
"Bug Tracker" = "https://github.com/kiku-jw/reliapi/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["reliapi*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
